Replace Specinfra::Command::Module::Ss with Iproute::Port
