--- lib/specinfra/command/module/iproute.rb
@@ -1,8 +1,0 @@
-module Specinfra
-  module Command
-    module Module
-      module Iproute
-      end
-    end
-  end
-end
