--- lib/specinfra/command/module/iproute/port.rb
@@ -1,16 +1,0 @@
-module Specinfra
-  module Command
-    module Module
-      module Iproute
-        module Port
-          def check_is_listening(port, options={})
-            pattern = ":#{port} "
-            pattern = " #{options[:local_address]}#{pattern}" if options[:local_address]
-            pattern = "^#{options[:protocol]} .*#{pattern}" if options[:protocol]
-            "ss -tunl | grep -- #{escape(pattern)}"
-          end
-        end
-      end
-    end
-  end
-end
