+++ lib/specinfra/command/module/ss.rb
@@ -1,0 +1,14 @@
+module Specinfra
+  module Command
+    module Module
+      module Ss
+        def check_is_listening(port, options={})
+          pattern = ":#{port} "
+          pattern = " #{options[:local_address]}#{pattern}" if options[:local_address]
+          pattern = "^#{options[:protocol]} .*#{pattern}" if options[:protocol]
+          "ss -tunl | grep -- #{escape(pattern)}"
+        end
+      end
+    end
+  end
+end
