+++ lib/specinfra/command/redhat/v7/port.rb
@@ -1,5 +1,5 @@
 class Specinfra::Command::Redhat::V7::Port < Specinfra::Command::Redhat::Base::Port
   class << self
-    include Specinfra::Command::Module::Iproute::Port
+    include Specinfra::Command::Module::Ss
   end
 end
