--- spec/command/module/iproute_spec.rb
@@ -1,4 +1,0 @@
-require 'spec_helper'
-describe Specinfra::Command::Module::Iproute do
-  let(:klass) { Specinfra::Command::Module::Iproute }
-end
